Extract FormattingFeature from EditorEventManager
